training_experiments/requirements.txt
# Training Experiments Dependencies
# Scripts for downloading datasets and training models

# Dataset download
kagglehub>=0.2.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Training utilities
albumentations>=1.3.0
tqdm>=4.66.0
tensorboard>=2.15.0

# Model conversion
onnx>=1.15.0
onnxruntime>=1.16.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
